% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{enable_react_debug_mode}
\alias{enable_react_debug_mode}
\title{Sets shiny.react into DEBUG mode.}
\usage{
enable_react_debug_mode()
}
\description{
Sets the \code{shiny.react_DEBUG} option to \code{value}. In DEBUG mode, shiny.react will load a dev version of React,
which is useful for debugging. It will also set a DEBUG logging level and pretty print tags sent to client.
}
