% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{enableReactDebugMode}
\alias{enableReactDebugMode}
\title{Enable 'React' debug mode}
\usage{
enableReactDebugMode()
}
\value{
Nothing. This function is called for its side effects.
}
\description{
Sets the \code{shiny.react_DEBUG} option to \code{TRUE}. In debug mode, 'shiny.react' will load a dev
version of 'React', which is useful for debugging. It will also set the logging level to DEBUG.
Call this function before running the app to enable the debugging mode.
}
