% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{html_dependency_react}
\alias{html_dependency_react}
\title{Shiny React dependency adding React libs.}
\usage{
html_dependency_react(use_cdn = FALSE, dev = FALSE)
}
\arguments{
\item{use_cdn}{If true, will load React from CDN instead of serving locally.}

\item{dev}{If true, will load a dev version of React (useful for debugging).}
}
\description{
Shiny React dependency adding React libs.
}
