% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{input_class_name}
\alias{input_class_name}
\title{Returns a CSS class name used to select DOM elements when installing Shiny input bindings.
It must be indentical to the name returned by \code{inputClassName()} in \code{shiny-react.jsx}.}
\usage{
input_class_name(package_name, component_name)
}
\arguments{
\item{package_name}{Name of the R package where the component can be found}

\item{component_name}{Name of the component}
}
\description{
Returns a CSS class name used to select DOM elements when installing Shiny input bindings.
It must be indentical to the name returned by \code{inputClassName()} in \code{shiny-react.jsx}.
}
