% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\name{JS}
\alias{JS}
\title{Mark character strings as literal JavaScript code}
\usage{
JS(...)
}
\arguments{
\item{...}{Character vectors as the JavaScript source code
(all arguments will be pasted into one character string).}
}
\value{
The input character vector marked with a special class.
}
\description{
Copied verbatim from the htmlwidgets package
to avoid adding a dependency just for this single function.
}
