% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{make_input}
\alias{make_input}
\title{Make Shiny input}
\usage{
make_input(html_dependencies, package_name, component_name, default_value = "")
}
\arguments{
\item{html_dependencies}{Dependencies to be included with this input}

\item{package_name}{Name of the R package where the component can be found}

\item{component_name}{Name of the React component}

\item{default_value}{Initial value for the input}
}
\description{
A helper which can be used to define a function behaving
just as Shiny input functions do.
}
