% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_output}
\alias{make_output}
\title{Make Shiny output}
\usage{
make_output(html_dependencies, package_name, component_name, ...)
}
\arguments{
\item{html_dependencies}{htmltools dependencies to attach (if NULL dependencies will not be attached).}

\item{package_name}{Package identifier, components should be in \code{window[package_name]} on the client.}

\item{component_name}{Component name on the client: should be in \code{window[package_name][component_name]}.}

\item{...}{Arguments and children to be passed to the React component.}
}
\description{
A helper which can be used to define a function behaving
just as Shiny output functions do.
}
