% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsx.R
\name{mark_as_react_tag}
\alias{mark_as_react_tag}
\title{Mark as React tag}
\usage{
mark_as_react_tag(package_name, component)
}
\arguments{
\item{package_name}{Where to find the component on the client.}

\item{component}{Component's name}
}
\description{
Mark an object as a React tag to be rendered on the client using React.
The component is expected to be available in \code{window[package_name][componentName]} on the client.
}
