% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsx.R
\name{needs_shiny_component_wrapper}
\alias{needs_shiny_component_wrapper}
\title{Should return TRUE iff a tag should be wrapped in a ShinyComponentWrapper}
\usage{
needs_shiny_component_wrapper(tag)
}
\arguments{
\item{tag}{HTML tag.}
}
\description{
Should return TRUE iff a tag should be wrapped in a ShinyComponentWrapper
}
