% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsx.R
\name{prepare_tags_for_serialization}
\alias{prepare_tags_for_serialization}
\title{This function does several things needed before a tag tree can be transformed into JSON:
\enumerate{
\item Remove all html_dependency nodes.
\item Rename some attributes in React nodes to ensure they do not become controlled components.
\item Remove names so that lists serialize into arrays and not objects.
\item Add ShinyComponentWrapper around Shiny inputs and outputs
}}
\usage{
prepare_tags_for_serialization(tags, is_in_attribute = FALSE)
}
\arguments{
\item{tags}{tags to be processed.}

\item{is_in_attribute}{should be set to true if the processed tags are inside another components attribute}
}
\description{
This function does several things needed before a tag tree can be transformed into JSON:
\enumerate{
\item Remove all html_dependency nodes.
\item Rename some attributes in React nodes to ensure they do not become controlled components.
\item Remove names so that lists serialize into arrays and not objects.
\item Add ShinyComponentWrapper around Shiny inputs and outputs
}
}
