% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{ReactContext}
\alias{ReactContext}
\title{React context}
\usage{
ReactContext(...)
}
\arguments{
\item{...}{Children to render.}
}
\description{
Render children with React.
}
\examples{
if (interactive()) shinyApp(
  ui = shiny.react:::ReactContext(
    "This text is rendered by React"
  ),
  server = function(input, output) {}
)
}
