% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{reactDependency}
\alias{reactDependency}
\title{Shiny React dependency adding React libs.}
\usage{
reactDependency(useCdn = FALSE)
}
\arguments{
\item{useCdn}{If true, will load React from CDN instead of serving locally.}
}
\description{
Shiny React dependency adding React libs.
}
