% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\name{reactOutput}
\alias{reactOutput}
\title{React output}
\usage{
reactOutput(outputId)
}
\arguments{
\item{outputId}{Id that can be used to render React on the server}
}
\description{
Creates a Shiny output which can be used analogously to \code{shiny::uiOutput()} but preserves
React state on re-renders.
}
