% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidgets.R
\name{reactWidget}
\alias{reactWidget}
\title{Creates a Shiny React widget that can be then rendered on the client with React}
\usage{
reactWidget(...)
}
\arguments{
\item{...}{tags to be rendered}
}
\description{
Creates a Shiny React widget that can be then rendered on the client with React
}
