% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsx.R
\name{rename_attribs_to_uncontrolled}
\alias{rename_attribs_to_uncontrolled}
\title{Rename controlled attributes to uncontrolled}
\usage{
rename_attribs_to_uncontrolled(tag)
}
\arguments{
\item{tag}{HTML tag to be processed.}
}
\description{
Rename controlled attributes to uncontrolled
}
\details{
For certain tags, some attributes (e.g. \verb{<input value="">}) make a component controlled in React.
This happens even if they are inserted using dangerouslySetInnerHTML.
We want the components to be uncontrolled, because we observe values via Shiny bindings.
Therefore we need to rename the attributes, e.g. value -> defaultValue.
Otherwise the value cannot be changed by the user, because React resets it to the initial value.

See for details:
\itemize{
\item https://reactjs.org/docs/uncontrolled-components.html
\item https://goshakkk.name/controlled-vs-uncontrolled-inputs-react/
}
}
