% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\name{renderReact}
\alias{renderReact}
\title{Render React}
\usage{
renderReact(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression returning the HTML / React to render.}

\item{env}{Environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression?}
}
\description{
Renders HTML and/or React in outputs created with \code{reactOutput()} (analogously to
\code{shiny::renderUI()}).
}
