% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsx.R
\name{serialize_shiny_react_tags}
\alias{serialize_shiny_react_tags}
\title{Serialize \code{JS}-wrapped attributes directly into the representation to be executed on the client.}
\usage{
serialize_shiny_react_tags(tags, pretty_print = FALSE)
}
\arguments{
\item{tags}{tags to be processed.}

\item{pretty_print}{If TRUE, the tags representation will be sent to the client in a readable format.}
}
\description{
Serialize \code{JS}-wrapped attributes directly into the representation to be executed on the client.
}
