% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{set_react_debug_mode}
\alias{set_react_debug_mode}
\title{Sets shiny.react into DEBUG mode.}
\usage{
set_react_debug_mode(value = TRUE)
}
\arguments{
\item{value}{TRUE (default) to enable debug mode}
}
\description{
Sets the \code{shiny.react_DEBUG} option to \code{value}. In DEBUG mode, shiny.react will load a dev version of React,
which is useful for debugging.
}
