% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{ShinyComponentWrapper}
\alias{ShinyComponentWrapper}
\title{Shiny Component Wrapper}
\usage{
ShinyComponentWrapper(...)
}
\arguments{
\item{...}{Component to be wrapped.}
}
\description{
Use if your component isn't visible on the start, disappears and appears
to make sure inputs and outputs are bound.
}
