% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{shinyReactDependency}
\alias{shinyReactDependency}
\title{'shiny.react' JavaScript dependency}
\usage{
shinyReactDependency()
}
\value{
An \code{htmlDependency} object which can be used attach the JavaScript code
required by 'shiny.react'.
}
\description{
'shiny.react' JavaScript dependency
}
