% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\name{triggerEvent}
\alias{triggerEvent}
\title{Trigger event}
\usage{
triggerEvent(inputId)
}
\arguments{
\item{inputId}{'Shiny' input ID to trigger the event on.}
}
\value{
A \code{ReactData} object which can be passed as a prop to 'React' components.
}
\description{
Creates a handler which can be used for \code{onClick} and similar props of 'React' components
to trigger an event in 'Shiny'.
}
