% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{update_input}
\alias{update_input}
\title{Update Shiny input}
\usage{
update_input(session, input_id, value = NULL, ...)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer}

\item{input_id}{The id of the input object.}

\item{value}{The value to set for the input object. If NULL, value will be unchanged.}

\item{...}{Additional parameters for the input to be updated. If empty, parameters will be unchanged.}
}
\description{
A generic function used to update inputs defined with \link{make_input}.
}
