% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsx.R
\name{withReact}
\alias{withReact}
\title{Use arbitrary React components and props in R.}
\usage{
withReact(...)
}
\arguments{
\item{...}{One or more tag objects which can be a mix of Shiny (htmltools) and React tags.}
}
\description{
Wrap your tags with this function to enable using a mix of HTML and React tags.
It will create a \verb{<div>} and \verb{<script>} tags on the client. The script will render tags inside the div using React.
}
\details{
Attributes wrapped in \code{JS(...)} (originally from \code{htmlwidgets}) will be passed directly as JS code.
It will be executed when the tags representation is received on the client, so usually they should be functions.

This function is named in camelCase for consistency with Shiny naming convention.
}
