% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\name{asProps}
\alias{asProps}
\title{Parse arguments as props}
\usage{
asProps(...)
}
\arguments{
\item{...}{Arguments to prepare for passing as props to a React component}
}
\description{
Converts arguments to a list which can be passed as the \code{props} argument to \code{reactElement()}.
}
