% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{reactDependency}
\alias{reactDependency}
\title{'React' library dependency}
\usage{
reactDependency(useCdn = FALSE)
}
\arguments{
\item{useCdn}{If \code{TRUE}, 'React' will be loaded from a CDN instead of being served locally.}
}
\value{
An \code{htmlDependency} object which can be used to attach the 'React' library.
}
\description{
'React' library dependency
}
