% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\name{reactElement}
\alias{reactElement}
\title{Create a React element}
\usage{
reactElement(module, name, props, deps = NULL)
}
\arguments{
\item{module}{JavaScript module to import the component from.}

\item{name}{Name of the component.}

\item{props}{Props to pass to the component.}

\item{deps}{HTML dependencies to attach.}
}
\description{
Create a React element
}
