% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidgets.R
\name{reactOutput}
\alias{reactOutput}
\title{Shiny binding for shiny.react HTML widget}
\usage{
reactOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Dimensions for the widget}
}
\description{
Shiny binding for shiny.react HTML widget
}
