% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidgets.R
\name{renderReact}
\alias{renderReact}
\title{Shiny binding for shiny.react HTML widget}
\usage{
renderReact(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression to be rendered}

\item{env}{Environment to be used}

\item{quoted}{Is expression quoted?}
}
\description{
Shiny binding for shiny.react HTML widget
}
