% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\name{setInput}
\alias{setInput}
\title{Set input}
\usage{
setInput(inputId, argIdx = 1)
}
\arguments{
\item{inputId}{Shiny input ID to set the value on.}

\item{argIdx}{Index of the argument to use as value.}
}
\description{
Set input
}
