% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\name{setInput}
\alias{setInput}
\title{Set input}
\usage{
setInput(inputId, argIdx = 1, debounce = NULL, throttle = NULL)
}
\arguments{
\item{inputId}{'Shiny' input ID to set the value on.}

\item{argIdx}{Index of the argument to use as value.}

\item{debounce}{The number of milliseconds to delay.}

\item{throttle}{The number of milliseconds to throttle invocations to.}
}
\value{
A \code{ReactData} object which can be passed as a prop to 'React' components.
}
\description{
Creates a handler which can be used for \code{onChange} and similar props of 'React' components
to set the value of a 'Shiny' input to one of the arguments passed to the handler.
}
