% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\name{triggerEvent}
\alias{triggerEvent}
\title{Trigger event}
\usage{
triggerEvent(inputId)
}
\arguments{
\item{inputId}{Shiny input ID to trigger the event on.}
}
\description{
Creates a handler which can be used for \code{onClick} and similar props of React components
to trigger an event in Shiny.
}
