% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\name{updateReactInput}
\alias{updateReactInput}
\title{Update React input}
\usage{
updateReactInput(session = shiny::getDefaultReactiveDomain(), inputId, ...)
}
\arguments{
\item{session}{Session object passed to function given to shinyServer.}

\item{inputId}{Id of the input object.}

\item{...}{Props to modify.}
}
\description{
Updates inputs created with the help of \code{InputAdapter} function (part of the JavaScript
interface). Analogous to \code{shiny::updateX()} family of functions, but generic.
}
\details{
If you're creating a wrapper package for a React library, you'll probably want to provide
a dedicated update function for each input to imitate Shiny interface.
}
